+++ lib/providers/display_predictions_provider.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final displayPredictionsProvider = StateProvider<bool>((ref) {
+  return false;
+});